update test and riot api
