+++ services/discord/tests/index.test.ts
@@ -1,7 +1,7 @@
 import { main } from "../src/index";
 import postgres from "../src/api/postgres";
 import discord from "../src/api/discord";
-import redis from "../src/api/redis";
+import { Redis } from "../src/api/redis";
 
 jest.mock("../src/Logger", () => ({
   createLogger: jest.fn().mockReturnValue({
@@ -12,12 +12,12 @@ jest.mock("../src/Logger", () => ({
 jest.mock("../src/api/postgres");
 jest.mock("../src/api/discord");
 jest.mock("../src/api/redis");
-jest.mock("../src/DataInitialization");
+jest.mock("../src/data/TftSummoner");
 
 it("runs expected", async () => {
   await main();
 
   expect(postgres.connect).toBeCalled();
   expect(discord.connect).toBeCalled();
-  expect(redis.connect).toBeCalled();
+  expect(Redis.connect).toBeCalled();
 });
